local/monitor/requirements.txt
python-dotenv
twilio==8.1.0
